mini-projet/auth-service/requirements.txt
Flask==3.1.0
PyJWT
Flask-Cors
